+++ src/Model/Uninstall/Uninstall.php
@@ -28,6 +28,8 @@ class Uninstall extends Model {
      *
      * @uses delete_option()      → removes option by name
      * @uses delete_site_option() → removes a option by name
+     *
+     * @return void
      */
     public function removeAll() {